Add a new plugin in a separate module that will generate kotlin services and models from an OpenAPI specification (#17)
Add a new module that holds a plugin that generates Kotlin model and service code from an OpenAPI spec. Can be easily integrated into a build with minimum configuration. I put this in a separate module because it has different dependencies than the other ones.

A simple unit test and a functional integration test are both included.