Update to use build conventions plugins from ronin-gradle
Introduces ronin-gradle v2 as a dependency to this project, and updates all the build files to use its conventions.  Also moves to using sonar and a more simplified build strategy, and uses axion-release for versioning.  Updates various dependency versions, kotlin to 1.8.22, and spring-boot to 3.0.7.  Moves the catalog output to a different module.